상품 등록 예외처리 추가

+++ src/main/java/com/idle/shoppingmall/mapper/Product/ProductMapper.java
@@ -13,7 +13,7 @@ import java.util.List;
 @Mapper
 public interface ProductMapper {
 
-    Long saveProduct(@Param("Product")Product productDTO);
+    Integer saveProduct(@Param("Product")Product productDTO);
 
     Product findById(Long productId);
     List<Product> findAllProducts();
